backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow>=9.5.0
psycopg2-binary>=2.9.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
supabase==2.0.2
python-dotenv==1.0.0
setuptools>=65.0.0
